Move dateRanges
