+++ packages/lib-user/src/components/shared/Select/Select.spec.js
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { composeStory } from '@storybook/react'
 
-import dateRanges from '../BarChart/helpers/dateRanges'
+import dateRanges from '../../../utils/dateRanges'
 import Meta, { DateRanges } from './Select.stories'
 
 describe('components > shared > Select', function() {
