+++ packages/lib-user/src/components/shared/Select/Select.stories.js
@@ -1,7 +1,8 @@
 import { Box } from 'grommet'
+
 import Select from './Select.js'
 
-import dateRanges from '../BarChart/helpers/dateRanges.js'
+import dateRanges from '../../../utils/dateRanges.js'
 
 export default {
   title: 'Components/shared/Select',
